backend/requirements.txt
numpy==1.18.2
asgiref==3.2.7
channels==2.4.0
django_multiselectfield==0.1.12
Django==3.0.5
djangorestframework==3.11.0
matplotlib==3.2.1
django-uniauth==1.2.2
django-allauth==0.42.0
django-cors-headers==3.3.0
django-rest-auth==0.9.5
psycopg2==2.8.5
channels-redis==2.4.2